changed add button to modal
